requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Dataset and Data Processing
datasets>=2.14.0
huggingface-hub>=0.19.0

# Code Analysis
ast-comments>=1.0.0
networkx>=3.1
graphviz>=0.20.1
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4

# RAG and Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Evaluation Metrics
nltk>=3.8.1
rouge-score>=0.1.2
evaluate>=0.4.1

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging and Monitoring
wandb>=0.16.0
tensorboard>=2.15.0
